streamlit/pyproject.toml
[project]
name = "apexml-streamlit"
version = "0.1.0"
description = "ApexML-Lite Streamlit app"
requires-python = ">=3.11,<3.14"
dependencies = [
    "streamlit>=1.30.0",
    "snowflake-snowpark-python>=1.11.0",
    "altair==5.5.0",
    "plotly>=5.20.0",
    "cryptography>=42.0.0",
    "fastf1>=3.3.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
]

[tool.uv]
package = false
